Changed mind about initialiser, moved back to config
