+++ src/config.ts
@@ -1,7 +1,9 @@
 import { context } from '@opentelemetry/api'
-import { ResolvedTraceConfig } from './types'
+import { ResolvedTraceConfig, Trigger } from './types'
 
 const configSymbol = Symbol('Otel Workers Tracing Configuration')
+
+export type Initialiser = (env: Record<string, unknown>, trigger: Trigger) => ResolvedTraceConfig
 
 export function setConfig(config: ResolvedTraceConfig, ctx = context.active()) {
 	return ctx.setValue(configSymbol, config)
