+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './sampling'
 export * from './sdk'
 export * from './span'
-export type * from './config'
+export type * from './types'
